+++ ChangeLog.md
@@ -1,5 +1,7 @@
-# Revision history for orders
+# Revision history for connections
 
-## 0.0.1  -- YYYY-mm-dd
+## 0.0.3  -- 2020-02-17
 
-* First version. Released on an unsuspecting world.
+* `Data.Float` : add cmath utils
+* `Data.Connection.Ratio` : add rational connections
+* `Numeric.Prelude` : add numeric prelude